PR feedback: unit test, refactor, organize
