+++ index/server/pkg/util/filter_test.go
@@ -1,0 +1,114 @@
+package util
+
+import (
+	"reflect"
+	"testing"
+
+	indexSchema "github.com/devfile/registry-support/index/generator/schema"
+)
+
+func TestFilterDevfileArchitectures(t *testing.T) {
+
+	tests := []struct {
+		name      string
+		index     []indexSchema.Schema
+		archs     []string
+		wantIndex []indexSchema.Schema
+	}{
+		{
+			name: "one arch filter",
+			index: []indexSchema.Schema{
+				{
+					Name:          "devfileA",
+					Architectures: []string{"amd64", "arm64"},
+				},
+				{
+					Name:          "devfileB",
+					Architectures: []string{"amd64"},
+				},
+				{
+					Name: "devfileC",
+				},
+			},
+			archs: []string{"amd64"},
+			wantIndex: []indexSchema.Schema{
+				{
+					Name:          "devfileA",
+					Architectures: []string{"amd64", "arm64"},
+				},
+				{
+					Name:          "devfileB",
+					Architectures: []string{"amd64"},
+				},
+				{
+					Name: "devfileC",
+				},
+			},
+		},
+		{
+			name: "two arch filters",
+			index: []indexSchema.Schema{
+				{
+					Name:          "devfileA",
+					Architectures: []string{"amd64", "arm64"},
+				},
+				{
+					Name:          "devfileB",
+					Architectures: []string{"amd64"},
+				},
+				{
+					Name: "devfileC",
+				},
+			},
+			archs: []string{"amd64", "arm64"},
+			wantIndex: []indexSchema.Schema{
+				{
+					Name:          "devfileA",
+					Architectures: []string{"amd64", "arm64"},
+				},
+				{
+					Name: "devfileC",
+				},
+			},
+		},
+		{
+			name: "empty filters",
+			index: []indexSchema.Schema{
+				{
+					Name:          "devfileA",
+					Architectures: []string{"amd64", "arm64"},
+				},
+				{
+					Name:          "devfileB",
+					Architectures: []string{"amd64"},
+				},
+				{
+					Name: "devfileC",
+				},
+			},
+			archs: []string{},
+			wantIndex: []indexSchema.Schema{
+				{
+					Name:          "devfileA",
+					Architectures: []string{"amd64", "arm64"},
+				},
+				{
+					Name:          "devfileB",
+					Architectures: []string{"amd64"},
+				},
+				{
+					Name: "devfileC",
+				},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			gotIndex := FilterDevfileArchitectures(test.index, test.archs)
+			if !reflect.DeepEqual(gotIndex, test.wantIndex) {
+				t.Errorf("Got: %v, Expected: %v", gotIndex, test.wantIndex)
+			}
+		})
+	}
+}
